+++ packages/typespec-go/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
 
+## 0.10.5 (unreleased)
+
+### Features Added
+
+* Added support for specific `@clientOption` decorators.
+  * `monomorphicResponseFieldName` on operations that return a monomorphic response to customize the response field name.
+  * `omitSerdeMethods` on models to skip emitting un/marshalers.
+
 ## 0.10.4 (2026-03-26)
 
 ### Bugs Fixed